refactor(design-system): flatten components into ui/ dir
