+++ packages/design-system/src/components/ui/badge.stories.tsx
@@ -1,6 +1,6 @@
 import { Circle } from 'lucide-react';
 
-import { Badge } from './Badge';
+import { Badge } from './badge';
 
 import type { Meta, StoryObj } from '@storybook/react';
