+++ packages/design-system/src/components/ui/button.stories.tsx
@@ -1,6 +1,6 @@
 import { ChevronRight, Plus } from 'lucide-react';
 
-import { Button } from './Button';
+import { Button } from './button';
 
 import type { Meta, StoryObj } from '@storybook/react';
