+++ packages/design-system/src/index.ts
@@ -1,9 +1,9 @@
-export { Icon, type IconProps, type IconSize } from './components/Icon/Icon';
-export { Button, type ButtonProps, buttonVariants } from './components/Button/Button';
-export { IconButton, type IconButtonProps } from './components/IconButton/IconButton';
-export { Input, type InputProps, PasswordInput, type PasswordInputProps } from './components/Input/Input';
-export { Badge, type BadgeProps, badgeVariants } from './components/Badge/Badge';
-export { FilterBadge, type FilterBadgeProps, filterBadgeVariants } from './components/Badge/FilterBadge';
-export { Spinner, type SpinnerProps } from './components/Spinner/Spinner';
-export { Tooltip, TooltipContent, TooltipKbd, type TooltipKbdProps, TooltipProvider, TooltipTrigger } from './components/Tooltip/Tooltip';
+export { Icon, type IconProps, type IconSize } from './components/ui/icon';
+export { Button, type ButtonProps, buttonVariants } from './components/ui/button';
+export { IconButton, type IconButtonProps } from './components/ui/icon-button';
+export { Input, type InputProps, PasswordInput, type PasswordInputProps } from './components/ui/input';
+export { Badge, type BadgeProps, badgeVariants } from './components/ui/badge';
+export { FilterBadge, type FilterBadgeProps, filterBadgeVariants } from './components/ui/filter-badge';
+export { Spinner, type SpinnerProps } from './components/ui/spinner';
+export { Tooltip, TooltipContent, TooltipKbd, type TooltipKbdProps, TooltipProvider, TooltipTrigger } from './components/ui/tooltip';
 export { cn } from './lib/cn';
